FEAT: Basic game play mode

+++ Source/Modules/ModuleEditor.h
@@ -32,7 +32,7 @@ public:
 	bool IsSceneFocused() const;
 	void SetResourceOnInspector(const std::weak_ptr<Resource>& resource) const;
 	void RefreshInspector() const;
-	std::pair<int, int> GetAvailableRegion();
+	std::pair<float, float> GetAvailableRegion();
 
 private:
 	void CopyAnObject();
